Rusty Backup — New Image Formats & Windows Improvements

  This release is a large expansion of the image formats Rusty Backup can read, write, and browse, plus a major overhaul of the Windows installation
  and update experience.

  New image format support

  - Norton Ghost (.gho / .ghs) — full read, inspect, and browse support, including:
    - File-aware FAT and NTFS images (browse individual files and folders inside the image)
    - Raw sector-mode images
    - Multi-file spanned image sets
    - Compressed images (both standard and high compression)
    - Ghost 7.5 full-disk images
    - Password-protected images, decrypted automatically
    - Reassembly into bootable whole-disk images for export and restore
  - WinImage (.imz) — read and browse support, including password-protected archives
  - QCOW2 (QEMU) — read, create, and in-place edit
  - VMDK (VMware) — read, create, and edit, for both flat and sparse disks
  - Dynamic VHD — read, create, and in-place edit (joining the existing fixed-VHD support)
  - BasiliskII / SheepShaver HFV — create blank HFV volumes, expand/convert HFS volumes to flat HFV, and full backup/restore round-trips

  All new formats are wired into the GUI (Inspect, Browse, Export) and the rb-cli command-line tool, and can be used as both backup sources and export
  targets where applicable.

  Windows experience

  - One-click installer — a proper Windows installer with file associations, "Add/Remove Programs" registration, and detection of portable vs.
  installed copies. Built automatically per architecture in CI.
  - In-app updates — "Download & Install Update" downloads and replaces the app in place, no manual reinstall needed.
  - Smarter elevation — administrator rights are now requested only when you actually need raw physical-disk access, instead of every launch. A
  top-bar "Show Physical Devices" control (with shield icon) requests elevation on demand, and physical-drive options stay disabled until elevated.

  Quality & infrastructure

  - Codebase-wide lint cleanup to a zero-warning state, plus a pre-commit hook and upfront test runs in CI.
  - Dependency upgrades and Windows 32/64-bit build and test fixes.
  - New planning/reference docs for MiSTer core filesystem support and additional Unix filesystems.

+++ src/fs/btrfs.rs
@@ -1098,7 +1098,7 @@ impl<R: Read + Seek + Send> CompactBtrfsReader<R> {
         let inner = CompactStreamReader::new(reader, layout);
 
         // data_size: only allocated sectors require disk reads.
-        let data_size = total_allocated_sectors * sector_size as u64;
+        let data_size = total_allocated_sectors * sector_size;
         Ok((
             Self { inner },
             CompactResult {
